third_party/WebKit/LayoutTests/http/tests/serviceworker/controller-on-reload.html

<!DOCTYPE html>
<title>Service Worker: Controller on reload</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<body>
<script>
var t = async_test('controller is set upon reload after registration');
t.step(function() {
    var reloaded = false;
    var scope = 'resources/blank.html';
    navigator.serviceWorker.unregister(scope).then(
        doTest,
        unreached_rejection(t, 'Unregister should not fail')
    );

    function doTest() {
        with_iframe(scope, t.step_func(onIframeLoad));
    }

    function onIframeLoad(frame) {
        var w = frame.contentWindow;

        function onActive(event) {
           assert_equals(w.navigator.serviceWorker.controller, null,
                         'controller should be null until the document is reloaded');
           reloaded = true;
           w.location.reload();
        }

        function onRegister(worker) {
            worker.addEventListener('statechange', t.step_func(function(event) {
                if (event.target.state == 'activated')
                    onActive();
            }));
        }

        if (reloaded) {
            assert_true(w.navigator.serviceWorker.controller instanceof w.ServiceWorker,
                        'controller should be a ServiceWorker object upon reload');
            service_worker_unregister_and_done(t, scope);
            return;
        }

        w.navigator.serviceWorker.register(
            'worker-no-op.js', {scope: 'blank.html'}
        ).then(
            t.step_func(onRegister),
            unreached_rejection(t, 'Registration should succeed, but failed')
        );
    }
});
</script>
</body>
