third_party/WebKit/LayoutTests/virtual/serviceworker/http/tests/serviceworker/controller-on-reload-expected.txt


This is a testharness.js-based test.
PASS controller is set upon reload after registration 
Harness: the test ran to completion.
